requirements.txt
streamlit
rembg
Pillow
opencv-python-headless
gfpgan
basicsr
facexlib
torch
torchvision
numpy